anime/requirements.txt
torch
kornia
opencv_python
ffmpeg-python